final change today

+++ lib/services/items_api.dart
@@ -1,24 +1,24 @@
 import 'dart:convert';
 
 import 'package:http/http.dart' as http;
-
+import 'package:shared_preferences/shared_preferences.dart';
 import '../../models/item_model.dart';
 
 class ItemsApi {
-  static const String baseUrl = "http://10.0.2.2:5000";
+  static const String baseUrl = "https://ktxzz6b2-5000.inc1.devtunnels.ms";
 
   // Post a lost Item api call
   static Future<http.Response> postItem(
       Map<String, dynamic> postDetails) async {
-    /* 
-        post details format: { 
+    /*
+        post details format: {
             type, // integer 0-lost, 1-found
             title, // String
             description, // String
             location1, //  String
             location2, // String
             image // Blob
-        } 
+        }
      */
     try {
       final response = await http.post(
@@ -49,14 +49,14 @@ class ItemsApi {
   // edit a post by its id
   static Future<http.Response> editItem(
       int itemId, Map<String, dynamic> postDetails) async {
-    /* 
-        post details format: { 
+    /*
+        post details format: {
             title, // String
             description, // String
             location1, //  String
             location2, // String
             image // Blob
-        } 
+        }
      */
     try {
       final response = await http.put(
@@ -73,10 +73,19 @@ class ItemsApi {
   // delete an item
   static Future<http.Response> deleteItem(int itemId) async {
     ///  pass itemId as argument
-
+    // get the cookies from shared preferences
+    final prefs = await SharedPreferences.getInstance();
+    //hardcoded cookie
+    final cookie = 'session=eyJ1c2VyX2lkIjoyNSwidXNlcl9yb2xlIjoxfQ.Z_BIuA.OsGKJErVnzjgsK6SO4QfToFxFok; HttpOnly; Path=/';// prefs.getString('cookies');
+    if (cookie == null) {
+      return http.Response(
+          jsonEncode({"message": "No cookies found", "statusCode": 401}),
+          401);
+    }
     try {
       final response = await http.delete(
         Uri.parse("$baseUrl/items/$itemId"),
+        headers: {"Cookie": cookie},
       );
       return response; // response data has item details in json form
     } catch (e) {
@@ -131,10 +140,18 @@ class ItemsApi {
 
   // report an item
   static Future<http.Response> reportItem(int itemId) async {
-
+    // get the cookies from shared preferences
+    final prefs = await SharedPreferences.getInstance();
+    final cookie = 'session=eyJ1c2VyX2lkIjoyLCJ1c2VyX3JvbGUiOjB9.Z_A24w.FNsRREbaInK147ygNZrdIA6XsnM; HttpOnly; Path=/';//prefs.getString('cookies');
+    if (cookie == null) {
+      return http.Response(
+          jsonEncode({"message": "No cookies found", "statusCode": 401}),
+          401);
+    }
     try {
       final response = await http.put(
-        Uri.parse("$baseUrl/items/$itemId/report"),
+          Uri.parse("$baseUrl/items/$itemId/report"),
+          headers: {"Cookie":cookie}
       );
       return response; //  Report count in response body as JSON
     } catch (e) {
